cryptopayments/requirements.txt
aiocryptopay==0.2.7
aiogram==2.25.1
aiohttp==3.8.4
aioschedule==0.5.2
fastapi==0.95.1
gunicorn==20.1.0
loguru==0.7.0
pydantic==1.10.7
python-dotenv==1.0.0
requests==2.30.0
starlette==0.26.1
StrEnum==0.4.10
Telethon==1.28.5
tortoise-orm==0.19.3
uvicorn==0.22.0
asyncpg==0.27.0
